Factor out some code to messageworker.go
